requirements.txt
noise>=1.2.2
PySDL2>=0.9.6
